Intent: Strengthen cached import coverage coherence Operators involved: [Coherence, Recursivity] Affected invariants: [#8] Key changes: - Consolidated cached import scenarios into tests/test_import_utils.py with logical groupings - Removed redundant tests/test_cached_import.py module Expected risks/dissonances: Minimal; reorganized tests rely on shared fixtures and warnings behavior Metrics: C(t) unchanged; pytest tests/test_import_utils.py Equivalence map: Cached import API unchanged
